requirements.txt
beautifulsoup4==4.13.4

branca==0.8.1
curl_cffi==0.12.0
folium==0.20.0
frozendict==2.4.6
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
mkl-service==2.4.0
multitasking==0.0.11
peewee==3.18.2
proto-plus==1.26.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.1
pywin32==307
rsa==4.9.1
setuptools==72.1.0
soupsieve==2.7
streamlit_folium==0.25.0
uritemplate==4.2.0
websockets==15.0.1
wheel==0.45.1
xyzservices==2025.4.0
yfinance==0.2.65
